# Conflicts:
#	src/main/java/controllers/CandidatureController.java
#	src/main/java/models/Bourse.java
#	src/main/java/models/Candidature.java
#	src/main/java/models/ConnexionJDBC.java
#	src/main/java/models/Enseignant.java
#	src/main/java/models/Etudiant.java

+++ src/main/java/models/Etudiant.java
@@ -47,7 +47,7 @@ public class Etudiant {
         return numeroEtudiant;
     }
 
-    public void setNumeroEtudiant(double numeroEtudiant) {
+    public void setNumeroEtudiant(int numeroEtudiant) {
         this.numeroEtudiant = numeroEtudiant;
     }
 
@@ -106,3 +106,5 @@ public class Etudiant {
         }
     }
 }
+//Pour les étudiants qui demandent une bourse, on veut connaitre leur nom, prénom, numéro
+//étudiant, note moyenne du dernier semestre validé.
